wallet_cmd file rename
